Extended the README
Explained the rough overview of the functionality of the CLI tool. Also described the dev setup and external crates used. This should give devs a good starting point, when new to the project. The additional features part was not discussed with the team, but is a simple braindump of things I thought would be nice-to-have.