Feature/setup cargo (#1)
* Initial cargo setup

This is just the output generated by the 'cargo new' command.

* Added cargo dependencies

We are using the clap crate for command line parsing. The serde crate with the derive feature are used for serialization and deserialization. Concretly, serde_yaml to work with the YAML file format.

* Added a VSC devcontainer configuration

This allows development in a standardized environment. Additionally, devs do not need to install cargo locally.

* Extended the README

Explained the rough overview of the functionality of the CLI tool. Also described the dev setup and external crates used. This should give devs a good starting point, when new to the project. The additional features part was not discussed with the team, but is a simple braindump of things I thought would be nice-to-have.