Feature/cli argument parsing (#3)
* Added the derive feature to clap crate

This was originally forgotten in the Cargo configuration.

* Added command line parsing

A general sketch of the command line parsing with four subcommands: run (send the request to the server given a specification), template (generate a template specification), status (show the status of a request) and list (list all current requests). These subcommands were not discussed in the team and are just first braindump of me.